:sparkles: feat: Custom exception delete

+++ src/main/java/com/josefroes/desafio_crud_de_clientes/services/ClientService.java
@@ -6,9 +6,11 @@ import com.josefroes.desafio_crud_de_clientes.repositories.ClientRepository;
 import com.josefroes.desafio_crud_de_clientes.services.exceptions.ResourceNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -52,13 +54,16 @@ public class ClientService {
 
             return new ClientDTO(entity);
         }catch (EntityNotFoundException e){
-            throw new ResourceNotFoundException("Recurso Não encontrado.");
+            throw new ResourceNotFoundException("Recurso não encontrado.");
         }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado.");
+        }
+            repository.deleteById(id);
     }
 
     private void copyDtoToEntity(Client entity, ClientDTO dto){
